% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_select_input.R
\name{country_select_input}
\alias{country_select_input}
\title{Custom Select Input}
\usage{
country_select_input(id, choices)
}
\arguments{
\item{id}{An id for the selectInput function you want to provide.}

\item{choices}{The choices you want to provide for the selectInput.}
}
\description{
This function is used to make custom input selections for different regions as a substitute for shiny function selectInput.
}
