% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidcolsa-package.R
\docType{package}
\name{covidcolsa-package}
\alias{covidcolsa}
\alias{covidcolsa-package}
\title{covidcolsa: Covid-19 Shiny Dashboard for Colombia and South Africa}
\description{
Contains data and statistics in R shiny app that analyses the Covid-19 situation in Colombia and South Africa. Counts of tests, cases confirmed, recoveries and deaths are summarized and compared between the two regions. The app itself and its functionalities is made available through functions and data from the package 'covidcolsa'. There are three main datasets which include one with both countries data named 'covid_colsa' and the other two named 'DailyCOL' and 'DailyZAF' for daily counts in the regions.
}
\author{
\strong{Maintainer}: Rahul Bharadwaj Mysore Venkatesh \email{rahulbharadwaj97@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
