% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DailyCOL.R
\docType{data}
\name{DailyCOL}
\alias{DailyCOL}
\title{Daily Covid-19 Data for Colombia}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 271 rows and 6 columns.
}
\source{
The data comes from Guidotti, E., Ardia, D., (2020), "COVID-19 Data Hub", Journal of Open Source Software 5(51):2376, doi: 10.21105/joss.02376, through the R package "COVID19" and filtered for the country Colombia and South Africa.
}
\usage{
DailyCOL
}
\description{
This data contains 271 observations and 6 variables for Covid stats in Colombia from January 2020. The data is wrangled from the following original data source.
}
\details{
\strong{id} - Country Code

\strong{date} - Date of observation

\strong{dailytest} - Daily Count of tests for a date

\strong{dailyconfirm} - Daily Count of confirmed for a date

\strong{dailyrecover} - Daily Count of recovered for a date

\strong{dailydeaths} - Daily Count of deaths for a date
}
\keyword{datasets}
