% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyplot.R
\name{dailyplot}
\alias{dailyplot}
\title{Daily Count Plot}
\usage{
dailyplot(data)
}
\arguments{
\item{data}{A dataset which consists of daily counts of all attributes.}
}
\description{
This function plots daily counts of tests, confirmed cases, recoveries, and deaths for the given data that has variable names "dailytested", "dailyconfirmed", "dailyrecovered", and "dailydeaths".
}
